Move create classes to pipeline-api
