+++ blueocean-github-pipeline/src/main/java/io/jenkins/blueocean/blueocean_github_pipeline/GithubPipelineCreateRequest.java
@@ -12,10 +12,10 @@ import hudson.model.User;
 import io.jenkins.blueocean.commons.ErrorMessage;
 import io.jenkins.blueocean.commons.ServiceException;
 import io.jenkins.blueocean.rest.Reachable;
-import io.jenkins.blueocean.rest.impl.pipeline.AbstractPipelineCreateRequest;
-import io.jenkins.blueocean.rest.impl.pipeline.credential.BlueOceanCredentialsProvider;
-import io.jenkins.blueocean.rest.impl.pipeline.credential.BlueOceanDomainRequirement;
-import io.jenkins.blueocean.rest.impl.pipeline.credential.CredentialsUtils;
+import io.jenkins.blueocean.pipeline.creation.AbstractPipelineCreateRequest;
+import io.jenkins.blueocean.pipeline.credential.BlueOceanCredentialsProvider;
+import io.jenkins.blueocean.pipeline.credential.BlueOceanDomainRequirement;
+import io.jenkins.blueocean.pipeline.credential.CredentialsUtils;
 import io.jenkins.blueocean.rest.model.BluePipeline;
 import io.jenkins.blueocean.rest.model.BlueScmConfig;
 import jenkins.branch.CustomOrganizationFolderDescriptor;
