+++ blueocean-pipeline-api/src/main/java/io/jenkins/blueocean/pipeline/credential/BlueOceanDomainRequirement.java
@@ -1,8 +1,6 @@
-package io.jenkins.blueocean.rest.impl.pipeline.credential;
+package io.jenkins.blueocean.pipeline.credential;
 
 import com.cloudbees.plugins.credentials.domains.DomainRequirement;
-
-import static io.jenkins.blueocean.rest.impl.pipeline.credential.BlueOceanDomainSpecification.DOMAIN_SPECIFICATION;
 
 /**
  * This is BlueOcean specific {@link DomainRequirement}.
