+++ blueocean-pipeline-api/src/main/java/io/jenkins/blueocean/pipeline/credential/BlueOceanDomainSpecification.java
@@ -1,7 +1,8 @@
-package io.jenkins.blueocean.rest.impl.pipeline.credential;
+package io.jenkins.blueocean.pipeline.credential;
 
 import com.cloudbees.plugins.credentials.domains.DomainRequirement;
 import com.cloudbees.plugins.credentials.domains.DomainSpecification;
+import io.jenkins.blueocean.pipeline.credential.BlueOceanDomainRequirement;
 
 import javax.annotation.Nonnull;
